Add API request validation

+++ src/main/java/com/aeq/transformers/impl/app/service/TransformerService.java
@@ -1,5 +1,6 @@
 package com.aeq.transformers.impl.app.service;
 
+import com.aeq.transformers.impl.app.exception.TransformerNotFoundException;
 import com.aeq.transformers.impl.app.model.Transformer;
 import com.aeq.transformers.impl.app.repository.TransformerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -7,6 +8,8 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.Optional;
 
 
 @Service
@@ -20,15 +23,29 @@ public class TransformerService {
     }
 
     public void updateTransformer(Transformer transformer){
-        transformerRepo.save(transformer);
+        if(findTransformerById(transformer.getId()) != null)
+           transformerRepo.save(transformer);
     }
 
-    public void deleteTransformer(Transformer transformer){
-        transformerRepo.delete(transformer);
+    public Transformer findTransformerById(Long transformerId) {
+        Optional<Transformer> result = transformerRepo.findById(transformerId);
+        Transformer transformer = null;
+        try {
+            transformer = result.get();
+        } catch(NoSuchElementException exception) {
+            throw new TransformerNotFoundException("Transformer not found with id: "+transformerId);
+        }
+        return transformer;
+    }
+
+    public void deleteTransformer(Transformer transformer) {
+        if(findTransformerById(transformer.getId()) != null)
+           transformerRepo.delete(transformer);
     }
 
     public void deleteTransformerById(Long transformerId){
-        transformerRepo.deleteById(transformerId);
+        if(findTransformerById(transformerId) != null)
+           transformerRepo.deleteById(transformerId);
     }
 
     public List<Transformer> getAllTransformers() {
